Move landing page components to new folder components/landing_page
